feat(scoped-elements): add getScopedTagName function
